+++ CHANGELOG.md
@@ -46,6 +46,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `ProofListIndexProxy#getProof` and `ProofListIndexProxy#getRangeProof` to return
   `UncheckedListProof` instead of `ListProof`. The latter is renamed into `ListProofNode`
   and may be accessed through `UncheckedListProof#getRootProofNode` (#516)
+- `com.exonum.binding.common.proofs.map.MapEntry` renamed to `ByteStringMapEntry` and 
+  `com.exonum.binding.storage.indices.MapEntry` renamed to `GenericMapEntry` and moved to common 
+  module. Interface `MapEntry` added for convenient access to MapEntries.
 
 ### Removed
 - `Hashing#toHexString`. (#379)